requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pandas==2.2.2
google-cloud-bigquery==3.25.0
google-api-python-client==2.143.0
google-auth==2.33.0
